requirements.txt
fastapi
uvicorn[standard]
requests
redis
python-dotenv
